connector/requirements.txt
supervisor==4.1.0
kafka-python==2.0.1
elasticsearch==7.6.0
pymongo==3.10.1
